requirements.txt
certifi~=2020.12.5
cffi~=1.14.5
cryptography~=3.4.7
Django~=3.2.3
django-rest-knox~=4.1.0
djangorestframework~=3.12.4
psycopg2~=2.8.6
pycparser~=2.20
PyYAML~=5.4.1
